Fix parenthesis in sanitizeCoordinates()
The first if statement checks it either offset is undefined or zero and maybe we have changedTouches data.
However, in the original parenthesis if offsetX = 0 then the whole expression was true even if the changedTouches was undefined, that way an exception was thrown from the inside of the if block. This fixes this error.