Give ability to set outline and stroke width in setProperties.
Added new options to setProperties for setting the lineWidth for
various states: outline_width, stroke_width, hi_outline_width, and
hi_stroke_width.

Added hi_outline option to setProperties which adjusts the color of the
highlighted outline.

Changed parsing of properties in setProperties to allow false values.
This allows the user to disable the outline or stroke altogether.  I
personally disable stroke and make the outline lineWidth larger to give
a cleaner look.

Use lineJoin = “round” so the rectangle edges are slightly rounded.
They are more pronounced if the line widths are set high.

Changed math of creating the rectangles so it accounts for the user
adjust lineWidth properties.  Outline and stroke are now always
rendered inside the selection instead of straddling it.