Fixed geometry differences between selection and saved annotation
Removed the minus 1 calculation to width and height when a selection is
saved.

Use the drawShape function to draw selections so there isn’t different
drawing code between selecting a new annotation and drawing an existing
one.

Only use lineJoin = “round” for the outline.  A rounded stroke inside
the outline looks bad.