Use the current annotation in the POPUP_SHOWN event
Currently, when a popup is shown it is dependant on the popup.show() caller to include the current annotation in the method call. Neither OpenSeadragon.viewer nor OpenLayers.viewer do this. Instead they do:
```js
  this._popup.setAnnotation(annotation);
  this._place_popup();
  this._popup.show();
```

This changes the event trigger to include use the currentAnnotation in the popup_shown event, rather than the passed in annotation.